Refactor import paths for Country and RESTCountry interfaces to improve module structure
