+++ src/app/country/components/country-list/country-list.component.ts
@@ -1,5 +1,5 @@
 import { Component, input } from '@angular/core';
-import { Country } from '../interfaces/country.interface';
+import { Country } from '../../interfaces/country.interface';
 import { DecimalPipe } from '@angular/common';
 import { RouterLink } from '@angular/router';
