+++ src/app/country/pages/country-page/country-information/country-information.component.ts
@@ -1,5 +1,5 @@
 import { Component, computed, input } from '@angular/core';
-import { Country } from '../../../components/interfaces/country.interface';
+import { Country } from '../../../interfaces/country.interface';
 import { DecimalPipe } from '@angular/common';
 
 @Component({
