+++ src/app/country/services/country.service.ts
@@ -1,8 +1,8 @@
 import { HttpClient } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
-import { RESTCountry } from '../components/interfaces/rest-countries.interfaces';
+import { RESTCountry } from '../interfaces/rest-countries.interfaces';
 import { map, Observable, catchError, throwError, delay } from 'rxjs';
-import { Country } from '../components/interfaces/country.interface';
+import { Country } from '../interfaces/country.interface';
 import { CountryMapper } from '../mappers/country.mapper';
 
 const API_URL = 'https://restcountries.com/v3.1';
